+++ CHANGELOG.md
@@ -22,6 +22,15 @@
 
 ----------------------------------------------
 
+# 2.1.2 Release notes
+
+### Bugfixes
+
+* When adding a nullable column of type Float while other columns existed
+  already, the values of the new column would be non-null. This is now fixed.
+
+----------------------------------------------
+
 # 2.1.1 Release notes
 
 ### Internals